Added folder for integrating C++ plugins and widgets
Added a folder Plugins\ for adding new C++ plugins and widgets. A python wrapper must be added to use these plugins in the main python application. This will help us to reuse any C or C++ code that is already written. Currently added plugin includes a TransferFunctionEditor.dll which contains a QWidget for editing 1D transfer function as a polyline. And, a double slider that can be used for editing grayscale window (contrast / brightness) similar to used in Radiology applications.